chore: Don't copy generator .jar file to build-tools/
Gradle complains that when multiple tasks have the same output path, it
cannot decide accurately whether they need to run or not. The
copyJarToBuildTools task shared the same output path as the other
copyJarToBuildTools tasks in the Android runtime repo, so all of them
were always getting run.

Instead, we can leave the .jar file in the location where it is built,
and reference it from there in the outer project, instead of reaching
into the outer project's build tools folder.