chore: Upgrade Gradle wrapper to 7.6
We want to do this upgrade in the outer project as well.

I have not looked too much into the Gradle 8.x series but it appears
there are a lot of features used that are deprecated in 8.x, so for now
this just upgrades to the latest in the 7.x series.